Conflicts:
	.gitignore
	src/Core/SQLite Database/CCP Database/CCPCategory.m
	src/Mac Eve Tools.xcodeproj/project.pbxproj

+++ src/Core/Skills/SkillPair.m
@@ -51,7 +51,7 @@
 -(NSString*) description
 {
 	SkillTree *st = [[GlobalData sharedInstance]skillTree];
-	return [NSString stringWithFormat:@"%@ %@ %ld",[st skillForId:typeID], typeID,skillLevel];
+	return [NSString stringWithFormat:@"%@ %@ %ld", [st skillForId:typeID], typeID, (long) skillLevel];
 }
 
 -(NSComparisonResult) compare:(SkillPair*)rhs
